+++ lib/extensions/theme_extension.dart
@@ -1,0 +1,9 @@
+import 'package:flutter/material.dart';
+
+extension ThemeExtension on ThemeData {
+  static bool responsiveValues = true;
+
+  set responsiveSpacing(bool test){
+    responsiveValues = false;
+  }
+}