+++ lib/models/responsive_data.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/material.dart';
-import 'package:responsive_spacing/models/layout_columns.dart';
-import 'package:responsive_spacing/models/scaled_size.dart';
+import 'package:responsive_spacing/responsive_spacing.dart';
 
 class ResponsiveData {
   final ScaledSize margin;
@@ -9,19 +7,19 @@ class ResponsiveData {
   final ScaledSize body;
   final LayoutColumns layoutColumns;
 
-  ResponsiveData(BuildContext context)
-      : margin = ScaledMargin.find(MediaQuery.of(context).size.width),
-        padding = ScaledPadding.find(MediaQuery.of(context).size.width),
-        gutter = ScaledGutter.find(MediaQuery.of(context).size.width),
-        body = ScaledBody.find(MediaQuery.of(context).size.width),
-        layoutColumns = LayoutColumns.find(MediaQuery.of(context).size.width);
+  ResponsiveData(double width)
+      : margin = ResponsiveSpacing.globalMargin.find(width),
+        padding = ResponsiveSpacing.globalPadding.find(width),
+        gutter = ResponsiveSpacing.globalGutter.find(width),
+        body = ResponsiveSpacing.globalBody.find(width),
+        layoutColumns = ResponsiveSpacing.globalColumns.find(width);
 
-  ResponsiveData.standard()
-      : margin = ScaledMargin.standard(),
-        padding = ScaledPadding.standard(),
-        gutter = ScaledGutter.standard(),
-        body = ScaledBody.standard(),
-        layoutColumns = LayoutColumns.standard() {
+  ResponsiveData.fallback()
+      : margin = const ScaledMargin.fallback(),
+        padding = const ScaledPadding.fallback(),
+        gutter = const ScaledGutter.fallback(),
+        body = const ScaledBody.fallback(),
+        layoutColumns = const LayoutColumns.fallback() {
     // ignore: avoid_print
     print(
         "WARNING: Standard responsive scaling is used. This usually happens if you use don't use the ResponsiveScaffold or the wrong context.");