+++ lib/responsive_spacing.dart
@@ -1,5 +1,6 @@
 library responsive_spacing;
 
+export 'package:responsive_spacing/responsive_spacing_main.dart';
 export 'package:responsive_spacing/widgets/spacing.dart';
 export 'package:responsive_spacing/widgets/responsive_scaffold.dart';
 export 'package:responsive_spacing/widgets/responsive_card.dart';
@@ -7,3 +8,9 @@ export 'package:responsive_spacing/widgets/responsive_grid.dart';
 export 'package:responsive_spacing/extensions/scaled_size_extension.dart';
 export 'package:responsive_spacing/extensions/responsive_data_extension.dart';
 export 'package:responsive_spacing/models/responsive_data.dart';
+export 'package:responsive_spacing/models/breakpoints.dart';
+export 'package:responsive_spacing/models/scaled_size.dart';
+export 'package:responsive_spacing/models/responsive_collection.dart';
+export 'package:responsive_spacing/models/responsive_layout_columns.dart';
+export 'package:responsive_spacing/extensions/theme_extension.dart';
+